% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{sassy_fastx_next}
\alias{sassy_fastx_next}
\title{Get the next FASTA/FASTQ batch}
\usage{
sassy_fastx_next(iter)
}
\arguments{
\item{iter}{An iterator created by \code{sassy_fastx_iter()}.}
}
\value{
\code{NULL} at end of file, otherwise a \code{sassy_fastx_batch} list with
\code{id}, \code{seq}, and \code{qual} elements. \code{id} is an ALTREP character vector, while
\code{seq} and \code{qual} are ALTREP lists whose elements are raw ALTREP vectors.
}
\description{
Get the next FASTA/FASTQ batch
}
\examples{
fq <- tempfile(fileext = ".fastq")
writeLines(c("@r1", "ACGT", "+", "!!!!"), fq, useBytes = TRUE)
it <- sassy_fastx_iter(fq, batch_records = 1)
batch <- sassy_fastx_next(it)
length(batch$id)
}
